hirectl-backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]

[project]
name = "hirectl"
version = "1.0.0"
description = "A real-time hiring intelligence system that identifies high-signal opportunities and tells you exactly how to act before they get saturated."
requires-python = ">=3.11"
dependencies = [
    # HTTP
    "httpx>=0.27",
    "aiohttp>=3.9",
    # Scraping
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "playwright>=1.52",
    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",
    # Caching / queuing
    "redis>=5.0",
    "rq>=1.16",
    # AI
    "openai>=1.30",
    "anthropic>=0.28",
    # Scheduling
    "apscheduler>=3.10",
    # Email
    "resend>=2.0",
    # Config
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "python-dotenv>=1.0",
    # API
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    # Utilities
    "rich>=13.7",
    "typer>=0.12",
    "tenacity>=8.3",
    "python-dateutil>=2.9",
    "scikit-learn>=1.5",
]

[project.scripts]
hirectl = "hirectl.cli:app"

[tool.setuptools.packages.find]
where = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
